Extract ConfigSource and tighten pipeline hook visibility
- Move ConfigSource trait and Env/Props/Args implementations into their
  own file; ConfigLoader now only owns merge + JSON parsing.
- AbstractPipelineApplication: keep main at the top of the class, define
  the abstract `run` and overridable `configure` at the bottom, and mark
  both as `protected` so they read as subclass extension points rather
  than public API.
- SparkPiDemo: align with parent visibility via `protected override`.
