+++ pipeline/src/main/scala/com/kakao/actionbase/pipeline/app/AbstractPipelineApplication.scala
@@ -7,11 +7,6 @@ import scala.reflect.ClassTag
 
 abstract class AbstractPipelineApplication[T <: Product: ClassTag] {
 
-  // Customization hook for the Spark session — e.g. enableHiveSupport() or
-  // .config("spark.sql.shuffle.partitions", "200"). Default is a no-op so
-  // each application opts in to whatever it actually needs.
-  protected def configure(builder: SparkSession.Builder): SparkSession.Builder = builder
-
   def main(args: Array[String]): Unit = {
     println(s"Running ${getClass.getSimpleName}")
 
@@ -29,5 +24,7 @@ abstract class AbstractPipelineApplication[T <: Product: ClassTag] {
     }
   }
 
-  def run(spark: SparkSession, config: T): Unit
+  protected def run(spark: SparkSession, config: T): Unit
+
+  protected def configure(builder: SparkSession.Builder): SparkSession.Builder = builder
 }
